Update .upptimerc.yml
External stylesheet reference is non-functional due to CORB policy and benfoster.dev server being inaccessible at this time. CSS has been changed to inline instead of external stylesheet reference. Google Fonts link has been established as missed in the previous commit. A misunderstanding of the documentation also caused the title of the page to be misinterpreted as the name setting under status-website.